pyproject.toml
[project]
name = "job-finder-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=3.2.1",
    "openai>=1.50.0,<1.100.0",
    "openai-agents>=0.2.8",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
